+++ src/m_cond.h
@@ -82,7 +82,7 @@ typedef struct
 	INT16 z;         ///< Z coordinate.
 	INT16 level;     ///< Level on which this emblem can be found.
 	UINT8 sprite;    ///< emblem sprite to use, 0 - 25
-	UINT8 color;     ///< skincolor to use
+	UINT16 color;     ///< skincolor to use
 	INT32 var;       ///< If needed, specifies information on the target amount to achieve (or target skin)
 	char hint[110];  ///< Hint for emblem hints menu
 	UINT8 collected; ///< Do you have this emblem?
@@ -93,7 +93,7 @@ typedef struct
 	char description[40];///< Description of goal (used in statistics)
 	UINT8 conditionset;  ///< Condition set that awards this emblem.
 	UINT8 sprite;        ///< emblem sprite to use, 0 - 25
-	UINT8 color;         ///< skincolor to use
+	UINT16 color;         ///< skincolor to use
 	UINT8 collected;     ///< Do you have this emblem?
 } extraemblem_t;
